Move compatibility into its own file
